fixed tests

+++ src/ml/reg/extended/w_xgboost.py
@@ -14,17 +14,11 @@ class Xgboost(XGB):
             d_valid = xgb.DMatrix(data_val, target_val)
         watchlist = [(d_train, 'train'), (d_valid, 'valid')]
         nrounds = num_steps
-        xgb_model = xgb.train(model_params, d_train, nrounds, watchlist, early_stopping_rounds=nrounds/2,
+        xgb_model = xgb.train(model_params, d_train, nrounds, watchlist, early_stopping_rounds=int(nrounds/2),
                               feval=obj_fn, maximize=True, verbose_eval=100)
         return self.ml_model(xgb, bst=xgb_model)
 
     def feature_importance(self):
-        #import pandas as pd
-        #gain = self.bst.feature_importance('gain')
-        #df = pd.DataFrame({'feature':self.bst.feature_name(), 
-        #    'split':self.bst.feature_importance('split'), 
-        #    'gain':100 * gain /gain.sum()}).sort_values('gain', ascending=False)
-        #return df
         pass
